feat(confidential-ledger): workign server, read data from guid
