+++ confidential-ledger-demo/server/confidentialledger.py
@@ -1,224 +1,90 @@
-# Copyright (c) 2021 Microsoft.
-# Copyright (c) 2021 Avanade Inc.
+# Import the Azure authentication library
 
+from azure.identity import DefaultAzureCredential
 
-from enum import Enum
+# import the data plane sdk
 
-from typing import Dict, NamedTuple, Iterable
-import argparse
+from azure.confidentialledger import ConfidentialLedgerClient
+from azure.confidentialledger.identity_service import (
+    ConfidentialLedgerIdentityServiceClient,
+)
 
-from dotenv import load_dotenv
-import requests
 import os
 import json
-
-from dataclasses_serialization.json import JSONSerializer
-
-
-class Certificates(NamedTuple):
-    """Collection of certificate filenames."""
-
-    network_cert_filename: str
-    public_key_filename: str
-    private_key_filename: str
+from dotenv import load_dotenv
 
 
-class AppData(NamedTuple):
-    """Collection of data required by the client."""
+def get_ledger_creds():
 
-    certificates: Certificates
-    receipts: Dict
-    identity_url: str
-    ledger_url: str
+    load_dotenv()
+    TENANT_ID = os.getenv("AZURE_TENANT_ID")
+    os.environ["AZURE_TENANT_ID"] = TENANT_ID
+    credential = DefaultAzureCredential()
 
+    resource_group = os.getenv("RESOURCE_GROUP")
+    ledger_name = os.getenv("RESOURCE_NAME")
+    subscription_id = os.getenv("AZURE_TENANT_ID")
 
-class Response:
-    """Base class for response objects."""
+    identity_url = "https://identity.accledger.azure.com"
+    ledger_url = "https://" + ledger_name + ".confidential-ledger.azure.com"
 
-    @staticmethod
-    # Microsoft TOxDO: this should be moved to a common dir as shared between different projects
-    def _validate_required_fields(json_response: Dict, required_fields: Iterable[str]):
-        for field in required_fields:
-            if field not in json_response:
-                raise ValueError(f"JSON response missing required field {field}")
+    return credential, ledger_name, identity_url, ledger_url
 
 
-class GetLedgerDataResponse(Response):
-    """Translate the state response retrieved from a CCF node."""
+def get_ledger_client():
 
-    view: str
-    seqno: str
+    credential, ledger_name, identity_url, ledger_url = get_ledger_creds()
 
-    @staticmethod
-    def from_json_response(json_response: Dict):
-        """Produce a LedgerResponse from a JSON response."""
-        print(json_response)
-        GetLedgerDataResponse._validate_required_fields(json_response, ["commit_info"])
+    identity_client = ConfidentialLedgerIdentityServiceClient(identity_url)
+    network_identity = identity_client.get_ledger_identity(ledger_id=ledger_name)
 
-        return GetLedgerDataResponse(
-            str(json_response["commit_info"]["view"]),
-            str(json_response["commit_info"]["seq-no"]),
-        )
+    ledger_tls_cert_file_name = "networkcert.pem"
+    with open(ledger_tls_cert_file_name, "w") as cert_file:
+        cert_file.write(network_identity.ledger_tls_certificate)
 
-    # Serialize
-    def to_json(self):
-        return JSONSerializer.serialize(self)
-
-
-class GetLedgerDataResponseWithBody(GetLedgerDataResponse):
-    """Translate the state response retrieved from a CCF node."""
-
-    # response: LedgerResponse
-    body: str
-
-    @staticmethod
-    def from_json_response(json_response: Dict):
-        """Produce a LedgerResponse from a JSON response."""
-        print(json_response)
-        GetLedgerDataResponse._validate_required_fields(
-            json_response, ["body", "commit_info"]
-        )
-
-        return GetLedgerDataResponseWithBody(
-            view=str(json_response["commit_info"]["view"]),
-            seqno=str(json_response["commit_info"]["seq-no"]),
-            body=json_response["body"],
-        )
-
-
-class RequestType(Enum):
-    """Enum for possible HTTP request types."""
-
-    GET = 0
-    POST = 1
-
-
-def getLedgerUri():
-
-    load_dotenv()  # take environment variables from .env.
-
-    return (os.getenv("IDENTITY_SERVICE_URI"), os.getenv("LEDGER_URI"))
-
-
-def createAppData():
-
-    parser = argparse.ArgumentParser("Sends commands to a Confidential Ledger")
-
-    parser.add_argument(
-        "--public-key", help="user public key", default="user0_cert.pem"
+    ledger_client = ConfidentialLedgerClient(
+        endpoint=ledger_url,
+        credential=credential,
+        ledger_certificate_path=ledger_tls_cert_file_name,
     )
 
-    parser.add_argument(
-        "--private-key", help="user private key", default="user0_privk.pem"
-    )
-
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        help="enable verbose output",
-        action="store_true",
-        default=False,
-    )
-
-    args = parser.parse_args()
-
-    identity_url, ledger_url = getLedgerUri()
-
-    # Create AppDataObject
-    app_data = AppData(
-        Certificates("", args.public_key, args.private_key),
-        {},
-        identity_url,
-        ledger_url,
-    )
-
-    return app_data
+    return ledger_client
 
 
-def rpc_get_latest():
-    appData = createAppData()
-    result = _make_request__(
-        appData, "latest", data={"id": 1}  # this should (be?) commitId
-    )
-    # Format to ledger Response
-    response = GetLedgerDataResponseWithBody.from_json_response(json.loads(result))
-    return response
+def append_cl(data):
+    ledger_client = get_ledger_client()
+    append_result = ledger_client.append_to_ledger(entry_contents=str(data))
+    print(append_result.transaction_id)
+    check = ledger_client.get_transaction_status(append_result.transaction_id)
+    print(check.state)
+
+    return append_result.transaction_id
 
 
-def rpc_put(body, wait_for_commit=False):
-    # Using the same LedgerId
-    data = {"id": 1, "body": body}
-    print(data)
-    appData = createAppData()
-    response = _make_request__(
-        appData, "latest", data=data, request_type=RequestType.POST
-    )
-    return response
+def read_all():
+    ledger_client = get_ledger_client()
+    all = []
+    for entry in ledger_client.get_ledger_entries():
+        all.append([entry.contents, entry.transaction_id])
+
+    return all
 
 
-def _make_request__(
-    app_data: AppData,
-    method: str,
-    data: Dict = None,
-    request_type: RequestType = RequestType.GET,
-):
-    _session = (requests.Session(),)
-    # _network_cert_path = self._ledger_client_phase.network_cert_path TODO get this another way
-    """Makes a request to the network."""
-    headers = (
-        None
-        if request_type == RequestType.GET
-        else {"content-type": "application/json"}
-    )
-    url = f"https://{app_data.ledger_url}/app/{method}"
-    if request_type == RequestType.GET:
-        response = _session.get(
-            url,
-            headers=headers,
-            # verify=self._network_cert_path,
-            cert=(
-                app_data.certificates.public_key_filename,
-                app_data.certificates.private_key_filename,
-            ),
-            params=data,
-        )
-    elif request_type == RequestType.POST:
-        response = _session.post(
-            url,
-            headers=headers,
-            # verify=self._network_cert_path,
-            cert=(
-                app_data.certificates.public_key_filename,
-                app_data.certificates.private_key_filename,
-            ),
-            json=data,
-        )
-    else:
-        raise RuntimeError("Unsupported request type")
+def search_entries_guid(search_guid):
+    """Returns a list of entries with the relavant guid"""
+    returns = []
 
-    expected_commit_info = {
-        "x-ccf-tx-seqno": "seq-no",
-        "x-ccf-global-commit": "global-commit",
-        "x-ccf-tx-view": "view",
-    }
+    ledger_client = get_ledger_client()
+    entries = ledger_client.get_ledger_entries()
 
-    commit_info = {}
-    for header_key, print_key in expected_commit_info.items():
-        if header_key not in response.headers:
-            continue
+    for entry in entries:
+        try:
+            json_read = json.loads(str(entry.contents))
+            if (json_read["Meta"]["guid"]) == search_guid:
+                returns.append(entry.contents)
 
-        commit_info[print_key] = response.headers[header_key]
+        except:
+            # not a valid json
+            pass
 
-    # Check for { because the put request says it's application/json type but is
-    # just a bool.
-    result_data = {}
-    if response.text.find("{") != -1:
-        result_data = json.loads(response.text)
-    else:
-        result_data["response"] = response.text
-
-    result_data["commit_info"] = commit_info
-
-    result_output = json.dumps(result_data, indent=4, sort_keys=True)
-
-    return result_output
+    return returns
